Added webpack support for the plugin
