+++ code.js
@@ -1,714 +1,1 @@
-"use strict";
-// ============================================================================
-// NEPALI UX DATA PLUGIN - STATIC DATA GENERATION
-// ============================================================================
-// ============================================================================
-// ENHANCED DATA GENERATION MODULES
-// ============================================================================
-class NepaliNameGenerator {
-    static generateName(gender, language, options) {
-        // Static generation only
-        const firstNames = gender === "male"
-            ? this.maleNames
-            : gender === "female"
-                ? this.femaleNames
-                : [...this.maleNames, ...this.femaleNames];
-        const firstName = firstNames[Math.floor(Math.random() * firstNames.length)];
-        // If only first name is requested
-        if ((options === null || options === void 0 ? void 0 : options.nameType) === "first") {
-            if (language === "nepali") {
-                return this.convertToDevanagari(firstName);
-            }
-            return firstName;
-        }
-        // Generate full name (default behavior)
-        const surname = this.surnames[Math.floor(Math.random() * this.surnames.length)];
-        let name = firstName;
-        // Add middle name if requested
-        if (options === null || options === void 0 ? void 0 : options.includeMiddleName) {
-            const middleName = this.middleNames[Math.floor(Math.random() * this.middleNames.length)];
-            name = `${firstName} ${middleName}`;
-        }
-        // Add surname for full name
-        name = `${name} ${surname}`;
-        if (language === "nepali") {
-            return this.convertToDevanagari(name);
-        }
-        return name;
-    }
-    static isDevanagari(text) {
-        // Check if text contains Devanagari characters
-        return /[\u0900-\u097F]/.test(text);
-    }
-    static convertToDevanagari(name) {
-        return name
-            .split(" ")
-            .map((part) => this.devanagariMap[part] || part)
-            .join(" ");
-    }
-}
-NepaliNameGenerator.maleNames = [
-    "Sandeep",
-    "Rajesh",
-    "Suresh",
-    "Ramesh",
-    "Dinesh",
-    "Bikash",
-    "Prakash",
-    "Ashok",
-    "Suman",
-    "Rajan",
-    "Kamal",
-    "Nabin",
-    "Sagar",
-    "Dipesh",
-    "Mahesh",
-    "Binod",
-    "Arjun",
-    "Krishna",
-    "Shyam",
-    "Hari",
-    "Gopal",
-    "Mohan",
-    "Roshan",
-    "Anil",
-];
-NepaliNameGenerator.femaleNames = [
-    "Sunita",
-    "Kamala",
-    "Sita",
-    "Rita",
-    "Gita",
-    "Anita",
-    "Binita",
-    "Sabina",
-    "Purnima",
-    "Sangita",
-    "Sarita",
-    "Nirmala",
-    "Shanti",
-    "Radha",
-    "Maya",
-    "Devi",
-    "Sushila",
-    "Manju",
-    "Laxmi",
-    "Uma",
-    "Indira",
-    "Rekha",
-    "Sushma",
-    "Prema",
-];
-NepaliNameGenerator.surnames = [
-    "Shrestha",
-    "Gurung",
-    "Tamang",
-    "Rai",
-    "Limbu",
-    "Magar",
-    "Thapa",
-    "Sharma",
-    "Poudel",
-    "Adhikari",
-    "Khadka",
-    "Karki",
-    "Bhattarai",
-    "Neupane",
-    "Acharya",
-    "Koirala",
-    "Pandey",
-    "Gautam",
-    "Bhandari",
-    "Joshi",
-    "Ghimire",
-    "Rijal",
-    "Sapkota",
-    "Dahal",
-];
-NepaliNameGenerator.middleNames = [
-    "Kumar",
-    "Prasad",
-    "Bahadur",
-    "Lal",
-    "Das",
-    "Kumari",
-    "Raj",
-    "Man",
-    "Bir",
-    "Ratna",
-    "Chandra",
-    "Surya",
-    "Indra",
-    "Ram",
-    "Shiva",
-    "Vishnu",
-    "Mani",
-    "Prem",
-    "Ganga",
-    "Saraswati",
-    "Lakshmi",
-];
-NepaliNameGenerator.devanagariMap = {
-    // Male names
-    Sandeep: "सन्दीप",
-    Rajesh: "राजेश",
-    Suresh: "सुरेश",
-    Ramesh: "रमेश",
-    Dinesh: "दिनेश",
-    Bikash: "बिकाश",
-    Prakash: "प्रकाश",
-    Ashok: "अशोक",
-    Suman: "सुमन",
-    Rajan: "राजन",
-    Kamal: "कमल",
-    Nabin: "नबिन",
-    Sagar: "सागर",
-    Dipesh: "दिपेश",
-    Mahesh: "महेश",
-    Binod: "बिनोद",
-    Arjun: "अर्जुन",
-    Krishna: "कृष्ण",
-    Shyam: "श्याम",
-    Hari: "हरि",
-    Gopal: "गोपाल",
-    Mohan: "मोहन",
-    Roshan: "रोशन",
-    Anil: "अनिल",
-    // Female names
-    Sunita: "सुनिता",
-    Kamala: "कमला",
-    Sita: "सीता",
-    Rita: "रिता",
-    Gita: "गीता",
-    Anita: "अनिता",
-    Binita: "बिनिता",
-    Sabina: "सबिना",
-    Purnima: "पूर्णिमा",
-    Sangita: "संगीता",
-    Sarita: "सरिता",
-    Nirmala: "निर्मला",
-    Shanti: "शान्ति",
-    Radha: "राधा",
-    Maya: "माया",
-    Devi: "देवी",
-    Sushila: "सुशिला",
-    Manju: "मञ्जु",
-    Laxmi: "लक्ष्मी",
-    Uma: "उमा",
-    Indira: "इन्दिरा",
-    Rekha: "रेखा",
-    Sushma: "सुष्मा",
-    Prema: "प्रेमा",
-    // Surnames
-    Shrestha: "श्रेष्ठ",
-    Gurung: "गुरुङ",
-    Tamang: "तामाङ",
-    Rai: "राई",
-    Limbu: "लिम्बु",
-    Magar: "मगर",
-    Thapa: "थापा",
-    Sharma: "शर्मा",
-    Poudel: "पौडेल",
-    Adhikari: "अधिकारी",
-    Khadka: "खड्का",
-    Karki: "कार्की",
-    Bhattarai: "भट्टराई",
-    Neupane: "न्यौपाने",
-    Acharya: "आचार्य",
-    Koirala: "कोइराला",
-    Pandey: "पाण्डेय",
-    Gautam: "गौतम",
-    Bhandari: "भण्डारी",
-    Joshi: "जोशी",
-    Ghimire: "घिमिरे",
-    Rijal: "रिजाल",
-    Sapkota: "सापकोटा",
-    Dahal: "दाहाल",
-    // Middle names
-    Kumar: "कुमार",
-    Prasad: "प्रसाद",
-    Bahadur: "बहादुर",
-    Lal: "लाल",
-    Das: "दास",
-    Kumari: "कुमारी",
-    Raj: "राज",
-    Man: "मान",
-    Bir: "बीर",
-    Ratna: "रत्न",
-    Chandra: "चन्द्र",
-    Surya: "सूर्य",
-    Indra: "इन्द्र",
-    Ram: "राम",
-    Shiva: "शिव",
-    Vishnu: "विष्णु",
-    Mani: "मणि",
-    Prem: "प्रेम",
-    Ganga: "गंगा",
-    Saraswati: "सरस्वती",
-    Lakshmi: "लक्ष्मी",
-};
-class NepaliAddressGenerator {
-    static generateAddress(provinceFilter) {
-        // Static generation only
-        let availableProvinces = this.provinces;
-        if (provinceFilter) {
-            availableProvinces = this.provinces.filter((p) => p.name === provinceFilter);
-        }
-        const province = availableProvinces[Math.floor(Math.random() * availableProvinces.length)];
-        const district = province.districts[Math.floor(Math.random() * province.districts.length)];
-        const locations = this.locations[district] || ["Center"];
-        const location = locations[Math.floor(Math.random() * locations.length)];
-        // Generate house number
-        const houseNumber = Math.floor(Math.random() * 999) + 1;
-        // Get street name
-        const streetName = this.streetNames[Math.floor(Math.random() * this.streetNames.length)];
-        // Get postal code
-        const postalCodes = this.postalCodes[district] || ["44600"];
-        const postalCode = postalCodes[Math.floor(Math.random() * postalCodes.length)];
-        return `${houseNumber} ${streetName}, ${location}, ${district}, ${postalCode}, Nepal`;
-    }
-}
-NepaliAddressGenerator.provinces = [
-    {
-        name: "Koshi",
-        districts: [
-            "Jhapa",
-            "Ilam",
-            "Panchthar",
-            "Taplejung",
-            "Morang",
-            "Sunsari",
-        ],
-    },
-    {
-        name: "Madhesh",
-        districts: [
-            "Saptari",
-            "Siraha",
-            "Dhanusha",
-            "Mahottari",
-            "Sarlahi",
-            "Bara",
-        ],
-    },
-    {
-        name: "Bagmati",
-        districts: [
-            "Kathmandu",
-            "Lalitpur",
-            "Bhaktapur",
-            "Nuwakot",
-            "Rasuwa",
-            "Dhading",
-        ],
-    },
-    {
-        name: "Gandaki",
-        districts: ["Kaski", "Syangja", "Parbat", "Myagdi", "Baglung", "Mustang"],
-    },
-    {
-        name: "Lumbini",
-        districts: [
-            "Rupandehi",
-            "Kapilvastu",
-            "Dang",
-            "Banke",
-            "Bardiya",
-            "Gulmi",
-        ],
-    },
-    {
-        name: "Karnali",
-        districts: ["Surkhet", "Dailekh", "Jajarkot", "Kalikot", "Jumla", "Mugu"],
-    },
-    {
-        name: "Sudurpashchim",
-        districts: [
-            "Kailali",
-            "Kanchanpur",
-            "Doti",
-            "Bajhang",
-            "Bajura",
-            "Achham",
-        ],
-    },
-];
-NepaliAddressGenerator.streetNames = [
-    "Buddha Marg",
-    "King's Way",
-    "Durbar Marg",
-    "Kantipath",
-    "Ring Road",
-    "New Road",
-    "Kamal Pokhari",
-    "Maharajgunj Road",
-    "Lainchaur Road",
-    "Tripureshwor Road",
-    "Ratna Park Road",
-    "Sundhara Road",
-    "Bagbazaar Road",
-    "Dillibazaar Road",
-    "Thapathali Road",
-    "Kalimati Road",
-    "Gongabu Road",
-    "Balaju Road",
-    "Satdobato Road",
-    "Patan Road",
-];
-NepaliAddressGenerator.locations = {
-    Kathmandu: [
-        "Thamel",
-        "Baneshwor",
-        "New Road",
-        "Lazimpat",
-        "Maharajgunj",
-        "Baluwatar",
-        "Putalisadak",
-    ],
-    Lalitpur: [
-        "Patan Dhoka",
-        "Jawalakhel",
-        "Kupondole",
-        "Sanepa",
-        "Pulchowk",
-        "Mangal Bazaar",
-    ],
-    Bhaktapur: [
-        "Durbar Square",
-        "Pottery Square",
-        "Dattatraya Square",
-        "Taumadhi Square",
-    ],
-    Kaski: [
-        "Lakeside",
-        "Mahendrapul",
-        "Prithvi Chowk",
-        "Bagar",
-        "Damside",
-        "Bindyabasini",
-    ],
-    Jhapa: ["Birtamod", "Damak", "Bhadrapur", "Mechinagar", "Kankai"],
-    Morang: ["Biratnagar", "Urlabari", "Sundarharaicha", "Pathari"],
-};
-NepaliAddressGenerator.postalCodes = {
-    Kathmandu: ["44600", "44601", "44602", "44603", "44604", "44605"],
-    Lalitpur: ["44700", "44701", "44702", "44703"],
-    Bhaktapur: ["44800", "44801", "44802"],
-    Kaski: ["33700", "33701", "33702"],
-    Jhapa: ["57200", "57201", "57202"],
-    Morang: ["56600", "56601", "56602"],
-    Sunsari: ["56900", "56901"],
-    Ilam: ["57300", "57301"],
-    Rupandehi: ["32900", "32901"],
-    Chitwan: ["44200", "44201"],
-    Banke: ["21900", "21901"],
-};
-class NepaliContactGenerator {
-    static generatePhone(type) {
-        if (type === "mobile") {
-            const prefixes = ["98", "97", "96", "95", "94", "86", "85", "84"];
-            const prefix = prefixes[Math.floor(Math.random() * prefixes.length)];
-            const number = Math.floor(Math.random() * 10000000)
-                .toString()
-                .padStart(8, "0");
-            return `+977-${prefix}${number}`;
-        }
-        else {
-            const areaCodes = ["1", "21", "25", "61", "71", "81", "91"];
-            const areaCode = areaCodes[Math.floor(Math.random() * areaCodes.length)];
-            const number = Math.floor(Math.random() * 1000000)
-                .toString()
-                .padStart(7, "0");
-            return `+977-${areaCode}-${number}`;
-        }
-    }
-    static generateCurrency(amount) {
-        const value = amount || Math.floor(Math.random() * 100000) + 100;
-        const formatted = value.toLocaleString("en-IN");
-        return `Rs. ${formatted}`;
-    }
-    static generateDate(format) {
-        const date = new Date();
-        const adDate = date.toLocaleDateString("en-GB");
-        if (format === "ad")
-            return adDate;
-        if (format === "bs") {
-            const bsYear = date.getFullYear() + 57;
-            const bsMonth = Math.floor(Math.random() * 12) + 1;
-            const bsDay = Math.floor(Math.random() * 30) + 1;
-            return `${bsYear}/${bsMonth.toString().padStart(2, "0")}/${bsDay
-                .toString()
-                .padStart(2, "0")}`;
-        }
-        const bsYear = date.getFullYear() + 57;
-        return `${adDate} (BS: ${bsYear}/04/05)`;
-    }
-}
-class NepaliDateGenerator {
-    static convertToNepaliNumber(num) {
-        return num.toString().split('').map(digit => this.nepaliNumbers[parseInt(digit)]).join('');
-    }
-    static generateDate(language, format, includeTime) {
-        // Generate a random Nepali date (BS)
-        const currentYear = new Date().getFullYear();
-        const bsYear = currentYear + 56 + Math.floor(Math.random() * 3); // Current BS year ± some variation
-        const month = Math.floor(Math.random() * 12);
-        const day = Math.floor(Math.random() * 30) + 1;
-
-        //Generate a random time
-        const hours = Math.floor((Math.random() * 12) + 1); // Random hour from 1 to 12
-        const minutes = Math.floor(Math.random() * 60); // Random minute from 0 to 59
-        // const seconds = Math.floor(Math.random() * 60); // Random second from 0 to 59
-
-        // Pad single-digit numbers with a leading zero for formatting
-        // const formattedHours = String(hours).padStart(2, '0');
-        const formattedMinutes = String(minutes).padStart(2, '0');
-        // const formattedSeconds = String(seconds).padStart(2, '0');
-        const amPM = (Math.random() < 0.5 ? "AM" : "PM");
-
-        if (language === "nepali") {
-            const nepaliYear = this.convertToNepaliNumber(bsYear);
-            const nepaliMonth = this.convertToNepaliNumber(month + 1);
-            const nepaliDay = this.convertToNepaliNumber(day);
-            const nepaliHours = this.convertToNepaliNumber(hours);
-            const nepaliMinutes = this.convertToNepaliNumber(formattedMinutes);
-            const nepaliAmPM = amPM==="AM" ? "बिहानी" :  "बेलुकी"; 
-
-            if (format === "long") {
-                // Format: "२०७८ साल असर १५ गते"
-                return `${nepaliYear} साल ${this.nepaliMonths[month]} ${nepaliDay} गते` + ((includeTime) ? (`, ` + (`${nepaliAmPM} `) + (`${nepaliHours}:${nepaliMinutes}`)) : "");
-            }
-            else if (format === "short") {
-                // Format: "असार १५, २०७८"
-                return `${this.nepaliMonths[month]} ${nepaliDay}, ${nepaliYear}` + ((includeTime) ? (`, ` + (`${nepaliAmPM} `) + (`${nepaliHours}:${nepaliMinutes}`)) : "");
-            }
-            else {
-                // Format: "२०७८/०४/१५"
-                return `${nepaliYear}/${nepaliMonth.padStart(2, '०')}/${nepaliDay.padStart(2, '०')}` + ((includeTime) ? (`, ` + (`${nepaliAmPM} `) + (`${nepaliHours}:${nepaliMinutes}`)) : "");
-            }
-        }
-        else {
-            if (format === "long") {
-                // Format: "15 Ashar 2078 BS"
-                return (`${day} ${this.englishMonths[month]} ${bsYear} BS`) + ((includeTime) ? (`, ` + (`${hours}:${formattedMinutes} ${amPM}`)) : "");
-            }
-            else if (format === "short") {
-                // Format: "Ashar 15, 2078"
-                return `${this.englishMonths[month]} ${day}, ${bsYear}` + ((includeTime) ? (`, ` + (`${hours}:${formattedMinutes} ${amPM}`)) : "");
-            }
-            else {
-                // Format: "2078/04/15"
-                return `${bsYear}/${(month + 1).toString().padStart(2, '0')}/${day.toString().padStart(2, '0')}` + ((includeTime) ? (`, ` + (`${hours}:${formattedMinutes} ${amPM}`)) : "");
-            }
-        }
-    }
-}
-NepaliDateGenerator.nepaliMonths = [
-    "बैसाख", "जेठ", "असार", "साउन", "भदौ", "असोज",
-    "कार्तिक", "मंसिर", "पुष", "माघ", "फागुन", "चैत्र"
-];
-NepaliDateGenerator.englishMonths = [
-    "Baisakh", "Jestha", "Ashar", "Shrawan", "Bhadra", "Ashoj",
-    "Kartik", "Mangsir", "Poush", "Magh", "Falgun", "Chaitra"
-];
-NepaliDateGenerator.nepaliNumbers = ["०", "१", "२", "३", "४", "५", "६", "७", "८", "९"];
-// ============================================================================
-// FIGMA API INTEGRATION
-// ============================================================================
-class FigmaContentInserter {
-    static async insertText(content, nodes) {
-        const targetNodes = nodes || figma.currentPage.selection;
-        if (targetNodes.length === 0) {
-            const textNode = figma.createText();
-            await figma.loadFontAsync({ family: "Inter", style: "Regular" });
-            textNode.characters = content;
-            textNode.x = figma.viewport.center.x;
-            textNode.y = figma.viewport.center.y;
-            figma.currentPage.appendChild(textNode);
-            figma.currentPage.selection = [textNode];
-            figma.viewport.scrollAndZoomIntoView([textNode]);
-        }
-        else {
-            for (const node of targetNodes) {
-                if (node.type === "TEXT") {
-                    await figma.loadFontAsync(node.fontName);
-                    node.characters = content;
-                }
-            }
-        }
-    }
-    static async insertImage(imageUrl, nodes) {
-        try {
-            const targetNodes = nodes || figma.currentPage.selection;
-            if (targetNodes.length === 0) {
-                const rect = figma.createRectangle();
-                rect.resize(200, 200);
-                rect.x = figma.viewport.center.x;
-                rect.y = figma.viewport.center.y;
-                rect.fills = [{ type: "SOLID", color: { r: 0.8, g: 0.8, b: 0.8 } }];
-                figma.currentPage.appendChild(rect);
-                figma.currentPage.selection = [rect];
-                figma.viewport.scrollAndZoomIntoView([rect]);
-            }
-        }
-        catch (error) {
-            console.error("Failed to insert image:", error);
-        }
-    }
-    static async bulkInsert(contentArray, type) {
-        const selection = figma.currentPage.selection;
-        if (selection.length === 0) {
-            for (let i = 0; i < contentArray.length; i++) {
-                const content = contentArray[i];
-                if (type === "text") {
-                    const textNode = figma.createText();
-                    await figma.loadFontAsync({ family: "Inter", style: "Regular" });
-                    textNode.characters = content;
-                    textNode.x = figma.viewport.center.x;
-                    textNode.y = figma.viewport.center.y + i * 40;
-                    figma.currentPage.appendChild(textNode);
-                }
-            }
-            figma.currentPage.selection = figma.currentPage.children.slice(-contentArray.length);
-            figma.viewport.scrollAndZoomIntoView(figma.currentPage.selection);
-        }
-        else {
-            for (let i = 0; i < Math.min(selection.length, contentArray.length); i++) {
-                const node = selection[i];
-                const content = contentArray[i];
-                if (type === "text" && node.type === "TEXT") {
-                    await figma.loadFontAsync(node.fontName);
-                    node.characters = content;
-                }
-            }
-        }
-    }
-}
-// ============================================================================
-// MAIN PLUGIN LOGIC
-// ============================================================================
-figma.showUI(__html__, { width: 400, height: 600 });
-figma.ui.onmessage = async (msg) => {
-    try {
-        switch (msg.type) {
-            case "generate-name": {
-                const selection = figma.currentPage.selection;
-                const textNodes = selection.filter(node => node.type === "TEXT");
-                if (textNodes.length === 0) {
-                    // No text layers selected, create a new one
-                    const name = NepaliNameGenerator.generateName(msg.gender, msg.language, {
-                        includeMiddleName: msg.includeMiddleName,
-                        nameType: msg.nameType
-                    });
-                    await FigmaContentInserter.insertText(name);
-                }
-                else {
-                    // Update all selected text layers
-                    for (const textNode of textNodes) {
-                        const name = NepaliNameGenerator.generateName(msg.gender, msg.language, {
-                            includeMiddleName: msg.includeMiddleName,
-                            nameType: msg.nameType
-                        });
-                        await figma.loadFontAsync(textNode.fontName);
-                        textNode.characters = name;
-                    }
-                    figma.notify(`Updated ${textNodes.length} text layer${textNodes.length > 1 ? 's' : ''}`);
-                }
-                break;
-            }
-            case "generate-address": {
-                const selection = figma.currentPage.selection;
-                const textNodes = selection.filter(node => node.type === "TEXT");
-                if (textNodes.length === 0) {
-                    // No text layers selected, create a new one
-                    const address = NepaliAddressGenerator.generateAddress(msg.province);
-                    await FigmaContentInserter.insertText(address);
-                }
-                else {
-                    // Update all selected text layers
-                    for (const textNode of textNodes) {
-                        const address = NepaliAddressGenerator.generateAddress(msg.province);
-                        await figma.loadFontAsync(textNode.fontName);
-                        textNode.characters = address;
-                    }
-                    figma.notify(`Updated ${textNodes.length} text layer${textNodes.length > 1 ? 's' : ''}`);
-                }
-                break;
-            }
-            case "generate-nepali-date": {
-                const selection = figma.currentPage.selection;
-                const textNodes = selection.filter(node => node.type === "TEXT");
-                if (textNodes.length === 0) {
-                    // No text layers selected, create a new one
-                    const date = NepaliDateGenerator.generateDate(msg.language, msg.format);
-                    await FigmaContentInserter.insertText(date);
-                }
-                else {
-                    // Update all selected text layers
-                    for (const textNode of textNodes) {
-                        const date = NepaliDateGenerator.generateDate(msg.language, msg.format, msg.includeTime);
-                        await figma.loadFontAsync(textNode.fontName);
-                        textNode.characters = date;
-                    }
-                    figma.notify(`Updated ${textNodes.length} text layer${textNodes.length > 1 ? 's' : ''}`);
-                }
-                break;
-            }
-            case "generate-phone": {
-                const phone = NepaliContactGenerator.generatePhone(msg.phoneType);
-                if (msg.insert) {
-                    await FigmaContentInserter.insertText(phone);
-                }
-                else {
-                    figma.ui.postMessage({ type: "phone-generated", content: phone });
-                }
-                break;
-            }
-            case "generate-currency": {
-                const currency = NepaliContactGenerator.generateCurrency(msg.amount);
-                if (msg.insert) {
-                    await FigmaContentInserter.insertText(currency);
-                }
-                else {
-                    figma.ui.postMessage({
-                        type: "currency-generated",
-                        content: currency,
-                    });
-                }
-                break;
-            }
-            case "generate-date": {
-                const date = NepaliContactGenerator.generateDate(msg.format);
-                if (msg.insert) {
-                    await FigmaContentInserter.insertText(date);
-                }
-                else {
-                    figma.ui.postMessage({ type: "date-generated", content: date });
-                }
-                break;
-            }
-            case "insert-image":
-                await FigmaContentInserter.insertImage(msg.imageUrl || "");
-                break;
-            case "insert-text":
-                await FigmaContentInserter.insertText(msg.content || "");
-                break;
-            case "close-plugin":
-                figma.closePlugin();
-                break;
-        }
-    }
-    catch (error) {
-        console.error("Plugin error:", error);
-        figma.notify("An error occurred. Please try again.");
-    }
-};
-figma.on("selectionchange", () => {
-    const selection = figma.currentPage.selection;
-    const selectionInfo = {
-        count: selection.length,
-        types: selection.map((node) => node.type),
-        hasText: selection.some((node) => node.type === "TEXT"),
-    };
-    figma.ui.postMessage({ type: "selection-changed", selection: selectionInfo });
-});
+(()=>{"use strict";class a{static generateName(a,e,t){const n="male"===a?this.maleNames:"female"===a?this.femaleNames:[...this.maleNames,...this.femaleNames],r=n[Math.floor(Math.random()*n.length)];if("first"===(null==t?void 0:t.nameType))return"nepali"===e?this.convertToDevanagari(r):r;const i=this.surnames[Math.floor(Math.random()*this.surnames.length)];let s=r;return(null==t?void 0:t.includeMiddleName)&&(s=`${r} ${this.middleNames[Math.floor(Math.random()*this.middleNames.length)]}`),s=`${s} ${i}`,"nepali"===e?this.convertToDevanagari(s):s}static isDevanagari(a){return/[\u0900-\u097F]/.test(a)}static convertToDevanagari(a){return a.split(" ").map(a=>this.devanagariMap[a]||a).join(" ")}}a.maleNames=["Sandeep","Rajesh","Suresh","Ramesh","Dinesh","Bikash","Prakash","Ashok","Suman","Rajan","Kamal","Nabin","Sagar","Dipesh","Mahesh","Binod","Arjun","Krishna","Shyam","Hari","Gopal","Mohan","Roshan","Anil"],a.femaleNames=["Sunita","Kamala","Sita","Rita","Gita","Anita","Binita","Sabina","Purnima","Sangita","Sarita","Nirmala","Shanti","Radha","Maya","Devi","Sushila","Manju","Laxmi","Uma","Indira","Rekha","Sushma","Prema"],a.surnames=["Shrestha","Gurung","Tamang","Rai","Limbu","Magar","Thapa","Sharma","Poudel","Adhikari","Khadka","Karki","Bhattarai","Neupane","Acharya","Koirala","Pandey","Gautam","Bhandari","Joshi","Ghimire","Rijal","Sapkota","Dahal"],a.middleNames=["Kumar","Prasad","Bahadur","Lal","Das","Kumari","Raj","Man","Bir","Ratna","Chandra","Surya","Indra","Ram","Shiva","Vishnu","Mani","Prem","Ganga","Saraswati","Lakshmi"],a.devanagariMap={Sandeep:"सन्दीप",Rajesh:"राजेश",Suresh:"सुरेश",Ramesh:"रमेश",Dinesh:"दिनेश",Bikash:"बिकाश",Prakash:"प्रकाश",Ashok:"अशोक",Suman:"सुमन",Rajan:"राजन",Kamal:"कमल",Nabin:"नबिन",Sagar:"सागर",Dipesh:"दिपेश",Mahesh:"महेश",Binod:"बिनोद",Arjun:"अर्जुन",Krishna:"कृष्ण",Shyam:"श्याम",Hari:"हरि",Gopal:"गोपाल",Mohan:"मोहन",Roshan:"रोशन",Anil:"अनिल",Sunita:"सुनिता",Kamala:"कमला",Sita:"सीता",Rita:"रिता",Gita:"गीता",Anita:"अनिता",Binita:"बिनिता",Sabina:"सबिना",Purnima:"पूर्णिमा",Sangita:"संगीता",Sarita:"सरिता",Nirmala:"निर्मला",Shanti:"शान्ति",Radha:"राधा",Maya:"माया",Devi:"देवी",Sushila:"सुशिला",Manju:"मञ्जु",Laxmi:"लक्ष्मी",Uma:"उमा",Indira:"इन्दिरा",Rekha:"रेखा",Sushma:"सुष्मा",Prema:"प्रेमा",Shrestha:"श्रेष्ठ",Gurung:"गुरुङ",Tamang:"तामाङ",Rai:"राई",Limbu:"लिम्बु",Magar:"मगर",Thapa:"थापा",Sharma:"शर्मा",Poudel:"पौडेल",Adhikari:"अधिकारी",Khadka:"खड्का",Karki:"कार्की",Bhattarai:"भट्टराई",Neupane:"न्यौपाने",Acharya:"आचार्य",Koirala:"कोइराला",Pandey:"पाण्डेय",Gautam:"गौतम",Bhandari:"भण्डारी",Joshi:"जोशी",Ghimire:"घिमिरे",Rijal:"रिजाल",Sapkota:"सापकोटा",Dahal:"दाहाल",Kumar:"कुमार",Prasad:"प्रसाद",Bahadur:"बहादुर",Lal:"लाल",Das:"दास",Kumari:"कुमारी",Raj:"राज",Man:"मान",Bir:"बीर",Ratna:"रत्न",Chandra:"चन्द्र",Surya:"सूर्य",Indra:"इन्द्र",Ram:"राम",Shiva:"शिव",Vishnu:"विष्णु",Mani:"मणि",Prem:"प्रेम",Ganga:"गंगा",Saraswati:"सरस्वती",Lakshmi:"लक्ष्मी"};class e{static generateAddress(a){let e=this.provinces;a&&(e=this.provinces.filter(e=>e.name===a));const t=e[Math.floor(Math.random()*e.length)],n=t.districts[Math.floor(Math.random()*t.districts.length)],r=this.locations[n]||["Center"],i=r[Math.floor(Math.random()*r.length)],s=Math.floor(999*Math.random())+1,o=this.streetNames[Math.floor(Math.random()*this.streetNames.length)],h=this.postalCodes[n]||["44600"];return`${s} ${o}, ${i}, ${n}, ${h[Math.floor(Math.random()*h.length)]}, Nepal`}}e.provinces=[{name:"Koshi",districts:["Jhapa","Ilam","Panchthar","Taplejung","Morang","Sunsari"]},{name:"Madhesh",districts:["Saptari","Siraha","Dhanusha","Mahottari","Sarlahi","Bara"]},{name:"Bagmati",districts:["Kathmandu","Lalitpur","Bhaktapur","Nuwakot","Rasuwa","Dhading"]},{name:"Gandaki",districts:["Kaski","Syangja","Parbat","Myagdi","Baglung","Mustang"]},{name:"Lumbini",districts:["Rupandehi","Kapilvastu","Dang","Banke","Bardiya","Gulmi"]},{name:"Karnali",districts:["Surkhet","Dailekh","Jajarkot","Kalikot","Jumla","Mugu"]},{name:"Sudurpashchim",districts:["Kailali","Kanchanpur","Doti","Bajhang","Bajura","Achham"]}],e.streetNames=["Buddha Marg","King's Way","Durbar Marg","Kantipath","Ring Road","New Road","Kamal Pokhari","Maharajgunj Road","Lainchaur Road","Tripureshwor Road","Ratna Park Road","Sundhara Road","Bagbazaar Road","Dillibazaar Road","Thapathali Road","Kalimati Road","Gongabu Road","Balaju Road","Satdobato Road","Patan Road"],e.locations={Kathmandu:["Thamel","Baneshwor","New Road","Lazimpat","Maharajgunj","Baluwatar","Putalisadak"],Lalitpur:["Patan Dhoka","Jawalakhel","Kupondole","Sanepa","Pulchowk","Mangal Bazaar"],Bhaktapur:["Durbar Square","Pottery Square","Dattatraya Square","Taumadhi Square"],Kaski:["Lakeside","Mahendrapul","Prithvi Chowk","Bagar","Damside","Bindyabasini"],Jhapa:["Birtamod","Damak","Bhadrapur","Mechinagar","Kankai"],Morang:["Biratnagar","Urlabari","Sundarharaicha","Pathari"]},e.postalCodes={Kathmandu:["44600","44601","44602","44603","44604","44605"],Lalitpur:["44700","44701","44702","44703"],Bhaktapur:["44800","44801","44802"],Kaski:["33700","33701","33702"],Jhapa:["57200","57201","57202"],Morang:["56600","56601","56602"],Sunsari:["56900","56901"],Ilam:["57300","57301"],Rupandehi:["32900","32901"],Chitwan:["44200","44201"],Banke:["21900","21901"]};class t{static generatePhone(a){if("mobile"===a){const a=["98","97","96","95","94","86","85","84"];return`+977-${a[Math.floor(Math.random()*a.length)]}${Math.floor(1e7*Math.random()).toString().padStart(8,"0")}`}{const a=["1","21","25","61","71","81","91"];return`+977-${a[Math.floor(Math.random()*a.length)]}-${Math.floor(1e6*Math.random()).toString().padStart(7,"0")}`}}static generateCurrency(a){return`Rs. ${(a||Math.floor(1e5*Math.random())+100).toLocaleString("en-IN")}`}static generateDate(a){const e=new Date,t=e.toLocaleDateString("en-GB");if("ad"===a)return t;if("bs"===a){const a=e.getFullYear()+57,t=Math.floor(12*Math.random())+1,n=Math.floor(30*Math.random())+1;return`${a}/${t.toString().padStart(2,"0")}/${n.toString().padStart(2,"0")}`}return`${t} (BS: ${e.getFullYear()+57}/04/05)`}}class n{static convertToNepaliNumber(a){return a.toString().split("").map(a=>this.nepaliNumbers[parseInt(a)]).join("")}static generateDate(a,e,t){const n=(new Date).getFullYear()+56+Math.floor(3*Math.random()),r=Math.floor(12*Math.random()),i=Math.floor(30*Math.random())+1,s=Math.floor(12*Math.random()+1),o=Math.floor(60*Math.random()),h=String(o).padStart(2,"0"),l=Math.random()<.5?"AM":"PM";if("nepali"===a){const a=this.convertToNepaliNumber(n),o=this.convertToNepaliNumber(r+1),m=this.convertToNepaliNumber(i),g=this.convertToNepaliNumber(s),c=this.convertToNepaliNumber(h),d="AM"===l?"बिहानी":"बेलुकी";return"long"===e?`${a} साल ${this.nepaliMonths[r]} ${m} गते`+(t?`, ${d} ${g}:${c}`:""):"short"===e?`${this.nepaliMonths[r]} ${m}, ${a}`+(t?`, ${d} ${g}:${c}`:""):`${a}/${o.padStart(2,"०")}/${m.padStart(2,"०")}`+(t?`, ${d} ${g}:${c}`:"")}return"long"===e?`${i} ${this.englishMonths[r]} ${n} BS`+(t?`, ${s}:${h} ${l}`:""):"short"===e?`${this.englishMonths[r]} ${i}, ${n}`+(t?`, ${s}:${h} ${l}`:""):`${n}/${(r+1).toString().padStart(2,"0")}/${i.toString().padStart(2,"0")}`+(t?`, ${s}:${h} ${l}`:"")}}n.nepaliMonths=["बैसाख","जेठ","असार","साउन","भदौ","असोज","कार्तिक","मंसिर","पुष","माघ","फागुन","चैत्र"],n.englishMonths=["Baisakh","Jestha","Ashar","Shrawan","Bhadra","Ashoj","Kartik","Mangsir","Poush","Magh","Falgun","Chaitra"],n.nepaliNumbers=["०","१","२","३","४","५","६","७","८","९"];class r{static async insertText(a,e){const t=e||figma.currentPage.selection;if(0===t.length){const e=figma.createText();await figma.loadFontAsync({family:"Inter",style:"Regular"}),e.characters=a,e.x=figma.viewport.center.x,e.y=figma.viewport.center.y,figma.currentPage.appendChild(e),figma.currentPage.selection=[e],figma.viewport.scrollAndZoomIntoView([e])}else for(const e of t)"TEXT"===e.type&&(await figma.loadFontAsync(e.fontName),e.characters=a)}static async insertImage(a,e){try{if(0===(e||figma.currentPage.selection).length){const a=figma.createRectangle();a.resize(200,200),a.x=figma.viewport.center.x,a.y=figma.viewport.center.y,a.fills=[{type:"SOLID",color:{r:.8,g:.8,b:.8}}],figma.currentPage.appendChild(a),figma.currentPage.selection=[a],figma.viewport.scrollAndZoomIntoView([a])}}catch(a){console.error("Failed to insert image:",a)}}static async bulkInsert(a,e){const t=figma.currentPage.selection;if(0===t.length){for(let t=0;t<a.length;t++){const n=a[t];if("text"===e){const a=figma.createText();await figma.loadFontAsync({family:"Inter",style:"Regular"}),a.characters=n,a.x=figma.viewport.center.x,a.y=figma.viewport.center.y+40*t,figma.currentPage.appendChild(a)}}figma.currentPage.selection=figma.currentPage.children.slice(-a.length),figma.viewport.scrollAndZoomIntoView(figma.currentPage.selection)}else for(let n=0;n<Math.min(t.length,a.length);n++){const r=t[n],i=a[n];"text"===e&&"TEXT"===r.type&&(await figma.loadFontAsync(r.fontName),r.characters=i)}}}figma.showUI(__html__,{width:400,height:600}),figma.ui.onmessage=async i=>{try{switch(i.type){case"generate-name":{const e=figma.currentPage.selection.filter(a=>"TEXT"===a.type);if(0===e.length){const e=a.generateName(i.gender,i.language,{includeMiddleName:i.includeMiddleName,nameType:i.nameType});await r.insertText(e)}else{for(const t of e){const e=a.generateName(i.gender,i.language,{includeMiddleName:i.includeMiddleName,nameType:i.nameType});await figma.loadFontAsync(t.fontName),t.characters=e}figma.notify(`Updated ${e.length} text layer${e.length>1?"s":""}`)}break}case"generate-address":{const a=figma.currentPage.selection.filter(a=>"TEXT"===a.type);if(0===a.length){const a=e.generateAddress(i.province);await r.insertText(a)}else{for(const t of a){const a=e.generateAddress(i.province);await figma.loadFontAsync(t.fontName),t.characters=a}figma.notify(`Updated ${a.length} text layer${a.length>1?"s":""}`)}break}case"generate-nepali-date":{const a=figma.currentPage.selection.filter(a=>"TEXT"===a.type);if(0===a.length){const a=n.generateDate(i.language,i.format);await r.insertText(a)}else{for(const e of a){const a=n.generateDate(i.language,i.format,i.includeTime);await figma.loadFontAsync(e.fontName),e.characters=a}figma.notify(`Updated ${a.length} text layer${a.length>1?"s":""}`)}break}case"generate-phone":{const a=t.generatePhone(i.phoneType);i.insert?await r.insertText(a):figma.ui.postMessage({type:"phone-generated",content:a});break}case"generate-currency":{const a=t.generateCurrency(i.amount);i.insert?await r.insertText(a):figma.ui.postMessage({type:"currency-generated",content:a});break}case"generate-date":{const a=t.generateDate(i.format);i.insert?await r.insertText(a):figma.ui.postMessage({type:"date-generated",content:a});break}case"insert-image":await r.insertImage(i.imageUrl||"");break;case"insert-text":await r.insertText(i.content||"");break;case"close-plugin":figma.closePlugin()}}catch(a){console.error("Plugin error:",a),figma.notify("An error occurred. Please try again.")}},figma.on("selectionchange",()=>{const a=figma.currentPage.selection,e={count:a.length,types:a.map(a=>a.type),hasText:a.some(a=>"TEXT"===a.type)};figma.ui.postMessage({type:"selection-changed",selection:e})})})();
