+++ package.json
@@ -5,6 +5,8 @@
   "main": "code.js",
   "scripts": {
     "build": "tsc -p tsconfig.json",
+    "build-webpack": "webpack --config webpack.config.js",
+    "watch-webpack": "webpack --watch",
     "lint": "eslint --ext .ts,.tsx --ignore-pattern node_modules .",
     "lint:fix": "eslint --ext .ts,.tsx --ignore-pattern node_modules --fix .",
     "watch": "npm run build -- --watch"
@@ -16,8 +18,16 @@
     "@figma/plugin-typings": "^1.116.0",
     "@typescript-eslint/eslint-plugin": "^6.12.0",
     "@typescript-eslint/parser": "^6.12.0",
+    "css-loader": "^7.1.2",
     "eslint": "^8.54.0",
-    "typescript": "^5.3.2"
+    "html-inline-css-webpack-plugin": "^1.11.2",
+    "html-webpack-plugin": "^5.6.3",
+    "mini-css-extract-plugin": "^2.9.2",
+    "style-loader": "^4.0.0",
+    "typescript": "^5.3.2",
+    "webpack": "^5.101.0",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.2"
   },
   "eslintConfig": {
     "extends": [
